fix: Renamed overlay component to backdrop and added z-index
