+++ src/components/index.js
@@ -3,4 +3,4 @@ export { default as RahmetInput } from './RahmetInput/RahmetInput.vue';
 export { default as RahmetSpinner } from './RahmetSpinner/RahmetSpinner.vue';
 export { default as RahmetBottomSheet } from './RahmetBottomSheet/RahmetBottomSheet.vue';
 export { default as RahmetModal } from './RahmetModal/RahmetModal.vue';
-export { default as RahmetOverlay } from './RahmetOverlay/RahmetOverlay.vue';
+export { default as RahmetBackdrop } from './RahmetBackdrop/RahmetBackdrop.vue';
